+++ pareg_core/src/parsers.rs
@@ -356,35 +356,23 @@ pub fn split_arg<'a, T: FromArg<'a>>(
 /// struct Pair(i32, i32);
 ///
 /// impl FromRead for Pair {
-///     fn from_read(r: &mut pareg::Reader) -> ParseResult<Self> {
+///     fn from_read(r: &mut pareg::Reader) -> Result<(Self, Option<ArgError>)> {
 ///         let mut v = Pair::default();
-///         let r = parsef_part!(r, "({},{})", &mut v.0, &mut v.1);
-///         ParseResult::new(v, r)
+///         let r = parsef_part!(r, "({},{})", &mut v.0, &mut v.1)?;
+///         Ok((v, r))
 ///     }
 /// }
 ///
 /// assert_eq!(
 ///     arg_list::<Pair>("(1,2),(3,4),(5,6)", ",").unwrap(),
-///     vec![Pair(1, 2), Pair(3, 4), Pair(5, 6)],
+///     vec![Pair(1, 2), Pair(3, 4), Pair(5, 6)]
 /// );
 /// ```
 pub fn arg_list<T: FromRead>(arg: &str, sep: &str) -> Result<Vec<T>> {
     let mut res = vec![];
     let mut reader: Reader = arg.into();
     loop {
-        let start_pos = reader.pos().unwrap_or_default();
-        let item = T::from_read(&mut reader);
-        let Some(item) = item.res else {
-            return item.err.map_or_else(
-                || {
-                    reader
-                        .err_parse("Failed to parse.")
-                        .span_start(start_pos)
-                        .err()
-                },
-                |e| e.err(),
-            );
-        };
+        let (item, _) = T::from_read(&mut reader)?;
         res.push(item);
         if reader.peek()?.is_none() {
             return Ok(res);