+++ pareg_core/src/set_from_read.rs
@@ -6,15 +6,8 @@ pub trait SetFromRead {
 
 impl<T: FromRead> SetFromRead for T {
     fn set_from_read(&mut self, r: &mut Reader) -> Result<Option<ArgError>> {
-        let start = r.pos().unwrap_or_default();
-        let res = Self::from_read(r);
-        if let Some(v) = res.res {
-            *self = v;
-            Ok(res.err)
-        } else {
-            Err(res.err.unwrap_or_else(|| {
-                r.err_parse("Failed to parse argument.").span_start(start)
-            }))
-        }
+        let err;
+        (*self, err) = Self::from_read(r)?;
+        Ok(err)
     }
 }